Move never-consent code into subpath
